installer/templates/memory-v3/payload/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.11
requests==2.32.3
watchfiles==1.0.4
